+++ src/lib/dto/input-file.ts
@@ -1,3 +1,15 @@
-export class InputFile extends File {
+export class InputFile {
+    public id: any;
+    public file: File;
     public preview: string;
+
+    constructor(
+        id: any,
+        preview: string,
+        file?: File
+    ) {
+        this.id = id;
+        this.preview = preview;
+        this.file = file;
+    }
 }